app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
dotenv==0.9.9
et_xmlfile==2.0.0
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
frozenlist==1.6.0
fsspec==2025.3.2
gotrue==2.12.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pillow==11.2.1
pluggy==1.5.0
portalocker==2.10.1
postgrest==1.0.1
propcache==0.3.1
protobuf==6.31.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.25.5
pytesseract==0.3.13
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.2
realtime==2.4.3
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
sympy==1.14.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==14.2
Werkzeug==3.1.3
XlsxWriter==3.2.3
yarl==1.20.0
